requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiohttp>=3.9.0
sqlalchemy>=2.0.0
kiteconnect>=4.0.0
pytz>=2023.3

